+++ src/main/java/com/maryana/restspringboot/security/UserDetailsImpl.java
@@ -16,7 +16,7 @@ public class UserDetailsImpl implements UserDetails {
 
     private static final long serialVersionUID = 1L;
 
-    private int id;
+    private Long id;
 
     private String login;
 
@@ -25,7 +25,7 @@ public class UserDetailsImpl implements UserDetails {
 
     private Collection<? extends GrantedAuthority> authorities;
 
-    public UserDetailsImpl(int id, String username, String password,
+    public UserDetailsImpl(Long id, String username, String password,
                            Collection<? extends GrantedAuthority> authorities) {
         this.id = id;
         this.login = username;
@@ -50,7 +50,7 @@ public class UserDetailsImpl implements UserDetails {
         return authorities;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
